[fix] : Service를 지우고 Service를 UseCase로 변경
